dsp: targeted double precision for phase accumulators + filter integrators
Phase accumulators (BlepOsc, SynthLfo, SynthPhaser) and filter integrator
state (the Moog ladder z[], the SVF TPT state ic1/ic2) are now double, so
oscillator/LFO phase can't drift over long notes and high-resonance filters
stay numerically stable. Audio sample buffers stay float (no memory cost).
All 128 presets pass the audit and round-trip unchanged.

+++ src/dub/Svf.h
@@ -71,6 +71,6 @@ class Svf
   private:
     float sr_ = 48000.0f;
     float g_ = 0.1f, k_ = 2.0f, a1_ = 1.0f;
-    float ic1_ = 0.0f, ic2_ = 0.0f;
+    double ic1_ = 0.0, ic2_ = 0.0; // TPT integrator state — double for stability
 };
 } // namespace doobie
